Make resources directory and delete release specific directories
A release specific directory can be generated with sync-resources.sh -
it will copy modified files from the specified release if it exists
and copy the resources to the release.

That means you can modify the release's files and run
sync-resources.sh before committing. When pulling changes, run
sync-resources.sh again for the release you're working on.

Also removes generated files from the repository.